refactor table data
